pipeline/packages/gc-vision-adapter/pyproject.toml
[project]
name = "gc-vision-adapter"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.10.1",
    "google-cloud-vision>=3.13.0",
]

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
